GH-3914: Create a test utils for form components.
